fix bug that some databases that did not exist were trying to be deleted (Thanks to Pavel Fojt, owm.AT.fojt.cz for his report)
git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@579 9f51437f-89f6-0310-a5ff-ff9a3806d2ee

+++ src/cgi-bin/openwebmail/openwebmail-folder.pl
@@ -422,14 +422,18 @@ sub markreadfolder {
          }
       }
 
-      ow::dbm::unlinkdb($tmpdb) or
-         openwebmailerror(gettext('Cannot delete db:') . " $tmpdb ($!)");
+      if (ow::dbm::existdb($tmpdb)) {
+         ow::dbm::unlinkdb($tmpdb) or
+            openwebmailerror(gettext('Cannot delete db:') . " $tmpdb ($!)");
+      }
 
-      ow::filelock::lock($tmpfile, LOCK_UN) or
-         openwebmailerror(gettext('Cannot unlock file:') . " $tmpfile ($!)");
+      if (-f $tmpfile) {
+         ow::filelock::lock($tmpfile, LOCK_UN) or
+            openwebmailerror(gettext('Cannot unlock file:') . " $tmpfile ($!)");
 
-      unlink($tmpfile) or
-         openwebmailerror(gettext('Cannot delete file:') . " $tmpfile ($!)");
+         unlink($tmpfile) or
+            openwebmailerror(gettext('Cannot delete file:') . " $tmpfile ($!)");
+      }
 
       rmdir($tmpdir) or
          openwebmailerror(gettext('Cannot delete directory:') . " $tmpdir ($!)");
